Rename files containing classes to camel case.
