+++ lib/VeresOne.js
@@ -4,10 +4,10 @@
 'use strict';
 
 const constants = require('./constants');
-const VeresOneDidDoc = require('./veres-one-did-doc');
+const VeresOneDidDoc = require('./VeresOneDidDoc');
 const Injector = require('./Injector');
 const storage = require('./storage');
-const VeresOneClient = require('./veres-one-client');
+const VeresOneClient = require('./VeresOneClient');
 
 const DEFAULT_MODE = 'test';
