+++ lib/VeresOneClient.js
