+++ lib/VeresOneDidDoc.js
