+++ lib/index.js
@@ -16,7 +16,7 @@ injector.use('jsonld-signatures', jsigs);
 
 injector.env = {nodejs: true};
 
-const VeresOne = require('./veres-one');
+const VeresOne = require('./VeresOne');
 
 module.exports = {
   injector,
@@ -25,6 +25,6 @@ module.exports = {
   veres: (options) => {
     return new VeresOne({injector, ...options});
   },
-  VeresOneClient: require('./veres-one-client'),
-  VeresOneDidDoc: require('./veres-one-did-doc')
+  VeresOneClient: require('./VeresOneClient'),
+  VeresOneDidDoc: require('./VeresOneDidDoc')
 };
