+++ tests/VeresOneClient.spec.js
