+++ tests/VeresOneDidDoc.spec.js
